Fixed fetching bug
Fixed a bug where if you lose focus of the tab (alt-tab out and back in) it refetches the clubs. Also after the first search the placeholders were there at the same time as the previous search results. I fixed those and added a "No clubs found" when no clubs found